requirements.txt
# Snowflake dependencies
snowflake-connector-python>=3.0.0
snowflake-snowpark-python>=1.0.0
python-dotenv>=1.0.0

# Geospatial and satellite data
rioxarray==0.17.0
pystac==1.11.0
pystac_client==0.9.0
planetary_computer==1.0.0
odc-stac==0.3.10
shapely==2.1.1
rasterio==1.4.3
geopandas==1.1.1

# Data processing
pandas==2.3.0
xarray==2025.3.1
xarray[complete]
dask==2024.10.0
zarr==2.17.2
numcodecs==0.12.1
netCDF4==1.7.2
adlfs==2025.8.0

# Visualization
matplotlib==3.10.3
seaborn==0.13.2

# Machine Learning
scikit-learn==1.5.2

# Utilities
tqdm==4.66.5
numpy
scipy